# trie: [wip] reduce allocations in derivesha