--- a/trie/stacktrie.go
+++ b/trie/stacktrie.go
@@ -27,7 +27,7 @@ import (
 
 var (
 	stPool = sync.Pool{New: func() any { return new(stNode) }}
-	bPool  = newByteslicepool(32, 100)
+	bPool  = newBytesPool(32, 100)
 	_      = types.TrieHasher((*StackTrie)(nil))
 )
 
@@ -48,20 +48,19 @@ type StackTrie struct {
 	h          *hasher
 	last       []byte
 	onTrieNode OnTrieNode
-
-	keyScratch  []byte
-	pathScratch []byte
+	kBuf       []byte // buf space used for hex-key during insertions
+	pBuf       []byte // buf space used for path during insertions
 }
 
 // NewStackTrie allocates and initializes an empty trie. The committed nodes
 // will be discarded immediately if no callback is configured.
 func NewStackTrie(onTrieNode OnTrieNode) *StackTrie {
 	return &StackTrie{
-		root:        stPool.Get().(*stNode),
-		h:           newHasher(false),
-		onTrieNode:  onTrieNode,
-		keyScratch:  make([]byte, 0, 32),
-		pathScratch: make([]byte, 0, 32),
+		root:       stPool.Get().(*stNode),
+		h:          newHasher(false),
+		onTrieNode: onTrieNode,
+		kBuf:       make([]byte, 0, 32),
+		pBuf:       make([]byte, 0, 32),
 	}
 }
 
@@ -71,16 +70,14 @@ func (t *StackTrie) Update(key, value []byte) error {
 		return errors.New("trying to insert empty (deletion)")
 	}
 	var k []byte
-	{
-		// We can reuse the key scratch area, but only if the insert-method
-		// never holds on to it.
-		if cap(t.keyScratch) < 2*len(key) { // realloc to ensure sufficient cap
-			t.keyScratch = make([]byte, 2*len(key), 2*len(key))
+	{                                 // Need to expand the 'key' into hex-form. We use the dedicated buf for that.
+		if cap(t.kBuf) < 2*len(key) { // realloc to ensure sufficient cap
+			t.kBuf = make([]byte, 2*len(key), 2*len(key))
 		}
 		// resize to ensure correct size
-		t.keyScratch = t.keyScratch[:2*len(key)]
-		writeHexKey(t.keyScratch, key)
-		k = t.keyScratch
+		t.kBuf = t.kBuf[:2*len(key)]
+		writeHexKey(t.kBuf, key)
+		k = t.kBuf
 	}
 	if bytes.Compare(t.last, k) >= 0 {
 		return errors.New("non-ascending key order")
@@ -90,7 +87,7 @@ func (t *StackTrie) Update(key, value []byte) error {
 	} else {
 		t.last = append(t.last[:0], k...) // reuse key slice
 	}
-	t.insert(t.root, k, value, t.pathScratch[:0])
+	t.insert(t.root, k, value, t.pBuf[:0])
 	return nil
 }
 
@@ -173,9 +170,11 @@ func (n *stNode) getDiffIndex(key []byte) int {
 	return len(n.key)
 }
 
-// Helper function to that inserts a (key, value) pair into
-// the trie.
-// The key is not retained by this method, but always copied if needed.
+// Helper function to that inserts a (key, value) pair into the trie.
+//   - The key is not retained by this method, but always copied if needed.
+//   - The value is retained by this method, as long as the leaf that it represents
+//     remains unhashed. However: it is never modified.
+//   - The path is not retained by this method.
 func (t *StackTrie) insert(st *stNode, key, value []byte, path []byte) {
 	switch st.typ {
 	case branchNode: /* Branch */
@@ -408,7 +407,7 @@ func (t *StackTrie) hash(st *stNode, path []byte) {
 	// input values.
 	val := bPool.Get()
 	val = val[:32]
-	t.h.hashDataTo(blob, val)
+	t.h.hashDataTo(val, blob)
 	st.val = val
 
 	// Invoke the callback it's provided. Notably, the path and blob slices are