@rjl493456442 2025-09-02
```

[[ Master ]]
goos: darwin
goarch: arm64
pkg: github.com/ethereum/go-ethereum/core/types
cpu: Apple M1 Pro
BenchmarkDeriveSha200
BenchmarkDeriveSha200/std_trie
BenchmarkDeriveSha200/std_trie-8                6921        167578 ns/op       77865 B/op       1731 allocs/op
BenchmarkDeriveSha200/stack_trie
BenchmarkDeriveSha200/stack_trie-8              7146        168260 ns/op       26098 B/op        232 allocs/op

[[ PR ]]
goos: darwin
goarch: arm64
pkg: github.com/ethereum/go-ethereum/core/types
cpu: Apple M1 Pro
BenchmarkDeriveSha200
BenchmarkDeriveSha200/std_trie
BenchmarkDeriveSha200/std_trie-8                7058        172988 ns/op       80054 B/op       1926 allocs/op
BenchmarkDeriveSha200/stack_trie
BenchmarkDeriveSha200/stack_trie-8              7122        166626 ns/op         745 B/op         19 allocs/op
```