@holiman 2025-09-05
> I think you have 'master' and 'PR' reversed in your comment.

I don't think he had. Not sure which comment you referred to, but, first one: 

```
Master
BenchmarkDeriveSha200/stack_trie-8              7146        168260 ns/op       26098 B/op        232 allocs/op
PR
BenchmarkDeriveSha200/stack_trie-8              7122        166626 ns/op         745 B/op         19 allocs/op
```
Second one

```
Master
49936.95MB  1.11%  1.11% 68792.73MB  1.54%  github.com/ethereum/go-ethereum/core/types.encodeForDerive
   13106MB  0.29%  1.41% 17499.76MB  0.39%  github.com/ethereum/go-ethereum/core/types.Receipts.EncodeIndex

PR
14087.04MB  0.31%  0.31% 23043.54MB  0.51%  github.com/ethereum/go-ethereum/trie.(*StackTrie).Update
12947.95MB  0.29%   0.6% 17368.17MB  0.39%  github.com/ethereum/go-ethereum/core/types.Receipts.EncodeIndex
```
So `EncodeIndex` comparable, whereas pr has, as largest offender, `Update` at `14G` and master has `encodeForDerive` at `50G`.